fix CascasdeSelect functionality
the current versions of CascadeSelect return an array for the column definitions instead of a string. This causes htmlspechialchars to crash, which is called when constructing html fpr the form output. 
This fix chacks, if the definition is an array and sets the textinput to the name of the column and the default set in "(...)".
If the old version of CascadeSelect is used, the fix stay with the old behaviour.